# storage/api-docs/api-docs-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book App API",
        "description": "API documentation for the Book App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Base URL"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "8f20688904ee49d0d8739eeb08faf0b1",
                "requestBody": {
                    "$ref": "#/components/requestBodies/RegisterRequestBody"
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "token": {
                                                            "type": "string",
                                                            "example": "token-example"
                                                        },
                                                        "user": {
                                                            "$ref": "#/components/schemas/UserResource"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "985cd5203e2d0a820e113f2f4fdb2847",
                "requestBody": {
                    "$ref": "#/components/requestBodies/LoginRequestBody"
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "token": {
                                                            "type": "string",
                                                            "example": "token-example"
                                                        },
                                                        "user": {
                                                            "$ref": "#/components/schemas/UserResource"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/password/forgot": {
            "post": {
                "tags": [
                    "User Password"
                ],
                "summary": "Send a password reset link",
                "operationId": "b5a4d80ae9728ab520f234f73671354d",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ForgotPasswordRequestBody"
                },
                "responses": {
                    "200": {
                        "description": "Reset link sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/password/reset": {
            "post": {
                "tags": [
                    "User Password"
                ],
                "summary": "Reset the user password",
                "operationId": "a54c2cd3031d32dad026300d78e1cf7e",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ResetPasswordRequestBody"
                },
                "responses": {
                    "200": {
                        "description": "Password reset successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/password/update": {
            "put": {
                "tags": [
                    "User Password"
                ],
                "summary": "Update the authenticated user's password",
                "operationId": "52e791f68968132c6c82450f6e50ff48",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdatePasswordRequestBody"
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Current password is incorrect",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "An error occurred"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "description": "Represents a user in the system",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe@example.com"
                    },
                    "picture": {
                        "type": "string",
                        "example": "/storage/user/profile_pictures/1.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "ForgotPasswordRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "email"
                            ],
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "johndoe@example.com"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "LoginRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "email",
                                "password"
                            ],
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "johndoe@example.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "password123"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RegisterRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "email",
                                "password",
                                "password_confirmation"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "John Doe"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "johndoe@example.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "password123"
                                },
                                "password_confirmation": {
                                    "type": "string",
                                    "example": "password123"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ResetPasswordRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "email",
                                "password",
                                "password_confirmation",
                                "token"
                            ],
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "johndoe@example.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "newpassword123"
                                },
                                "password_confirmation": {
                                    "type": "string",
                                    "example": "newpassword123"
                                },
                                "token": {
                                    "type": "string",
                                    "example": "reset-token-example"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdatePasswordRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "current_password",
                                "new_password"
                            ],
                            "properties": {
                                "current_password": {
                                    "type": "string",
                                    "example": "oldpassword123"
                                },
                                "new_password": {
                                    "type": "string",
                                    "example": "newpassword123"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        }
    },
...
}